Add gnuplot warn

+++ examples/plot-simple.sh
@@ -13,6 +13,8 @@ STREAM=str1
 
 if ! xcompiler -q query-simple.rql ; then exit 1 ; fi
 
+if ! which gnuplot ; then echo "install gnuplot!" ; exit 1 ; fi
+
 rm nohup.out
 nohup xretractor &
